+++ src/control/app/src/main/kotlin/org/calypsonet/keyple/demo/control/data/model/ReaderType.kt
@@ -16,6 +16,5 @@ enum class ReaderType {
   BLUEBIRD,
   COPPERNIC,
   FAMOCO,
-  FLOWBIRD,
   NFC_TERMINAL
 }