+++ src/validation/app/src/main/kotlin/org/calypsonet/keyple/demo/validation/domain/model/ReaderType.kt
@@ -13,8 +13,8 @@
 package org.calypsonet.keyple.demo.validation.domain.model
 
 enum class ReaderType {
+  ARRIVE,
   BLUEBIRD,
   COPPERNIC,
-  FAMOCO,
-  FLOWBIRD
+  FAMOCO
 }